Move BookListItem into BookList directory
